libvirt-xen: Add a script attribute for disk specifications
Xen optionally allows specifying a script attribute when defining a
disk. This means that the specified device can actually be information
passed to the specified script. This allows users to create custom
devices which libxl and libvirt normally do not support. The script is
run on the backend domain, usually dom0 unless a driver domain is
specified.