libvirt-xen: Support memory hotplugging
The `<maxMemory>` domain XML element is about the maximum memory about
to be hot plugged. Before this patch Xen calculates the shadow memory
based on the maximum allocation (`<memory>`). In that case Xen only
allows up to twice that amount of memory to be hotplugged. This patch
adds support for the `<maxMemory>` element allowing one to specify the
maximum amount of memory to be hotplugged during configuration.