libvirt-xen: Allow marking backend domains as untrusted
When using backend domains, mainly with PCI passthrough, you may not
necessarily trust the backend. E.g. A NIC could be compromised by
malware in the driver domain. Xen allows marking a device as untrusted
and passing that information to the frontend. If the frontend supports
this attribute the frontend can take necessary precautions when handling
the device.

This patch does not change any default behavior, which is trusting the
backend domains.